[LEAD] refactor: atomic structure cleanup
